requirements.txt
# Sistema Insumos x Orçamento V4 - Dependências
# Core dependencies for production use

# HTTP requests and API client
requests==2.31.0

# Excel generation and manipulation
openpyxl==3.1.2

# Environment variables management
python-dotenv==1.0.0

# Data validation and settings management
pydantic==2.5.3
pydantic-settings==2.1.0

# SharePoint integration (Microsoft Graph API)
msal==1.24.1

# Date and time utilities (built-in datetime should be sufficient)
# But if needed for advanced date parsing:
# python-dateutil==2.8.2

# Optional: Enhanced JSON handling
# orjson==3.9.10

# Development and testing dependencies (optional)
# Uncomment if needed for development

# Testing framework
# pytest==7.4.3
# pytest-cov==4.1.0

# Code formatting and linting
# black==23.11.0
# flake8==6.1.0
# isort==5.12.0

# Type checking
# mypy==1.7.1

# Documentation generation
# mkdocs==1.5.3
# mkdocs-material==9.4.8

# Performance profiling
# memory-profiler==0.61.0
# line-profiler==4.1.1

# Enhanced logging (if needed beyond standard library)
# structlog==23.2.0

# Data analysis (if expanding functionality)
# pandas==2.1.3
# numpy==1.25.2

# Database connectivity (if adding DB support)
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.0          # MongoDB

# Security and authentication enhancements
# cryptography==41.0.7
# jwt==1.3.1

# Configuration management alternatives
# dynaconf==3.2.4
# hydra-core==1.3.2

# Async support (if needed for concurrent API calls)
# aiohttp==3.9.1
# asyncio-throttle==1.0.2

# Progress bars and CLI enhancements
# tqdm==4.66.1
# rich==13.7.0
# click==8.1.7

# Data serialization alternatives
# msgpack==1.0.7
# pickle5==0.0.12  # Python < 3.8 only

# Caching enhancements
# redis==5.0.1
# diskcache==5.6.3

# Monitoring and metrics
# prometheus-client==0.19.0
# sentry-sdk==1.38.0

# File format support extensions
# xlsxwriter==3.1.9   # Alternative Excel writer
# csv-diff==1.0        # CSV comparison
# tabulate==0.9.0      # Pretty table printing

# Backup and compression
# tarfile  # Built-in
# zipfile  # Built-in
# gzip     # Built-in

# Network and connectivity
# urllib3==2.1.0      # Usually comes with requests
# certifi==2023.11.17 # SSL certificates

# System information
# psutil==5.9.6       # System and process monitoring

# Minimum Python version requirement
# python_requires = ">=3.8"

# Platform-specific dependencies
# Windows-specific
# pywin32==306  # Windows only

# Linux-specific
# python-prctl==1.8.1  # Linux process control

# macOS-specific
# pyobjc-framework-Cocoa==10.0  # macOS only

# Notes:
# - All versions are pinned for stability
# - Comment/uncomment sections based on needs
# - Update versions periodically for security
# - Test thoroughly after version updates
# - Consider using pip-tools for dependency resolution